pyproject.toml
[tool.poetry]
name = "djangoreporttest"
version = "0.1.0"
description = ""
authors = ["daniilkimstach <daniilkimstachp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-rest-swagger = "^2.2.0"
drf-spectacular = "^0.27.2"
numpy = "^1.26.4"
openpyxl = "^3.1.2"
pandas = "^2.2.2"
xlsxwriter = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
